status screen updates
