+++ .gitignore
@@ -37,7 +37,7 @@ stats.html
 !.yarn/sdks
 !.yarn/versions
 yarn.lock
-interface/analyse.html
+analyse.html
 interface/vite.config.ts.timestamp*
 
 # scripts
