+++ interface/package.json
@@ -23,15 +23,15 @@
   },
   "dependencies": {
     "@alova/adapter-xhr": "^1.0.3",
-    "@babel/core": "^7.24.0",
+    "@babel/core": "^7.24.1",
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.0",
-    "@mui/icons-material": "^5.15.13",
-    "@mui/material": "^5.15.13",
+    "@mui/icons-material": "^5.15.14",
+    "@mui/material": "^5.15.14",
     "@table-library/react-table-library": "4.1.7",
     "@types/imagemin": "^8.0.5",
     "@types/lodash-es": "^4.17.12",
-    "@types/node": "^20.11.28",
+    "@types/node": "^20.11.30",
     "@types/react": "^18.2.67",
     "@types/react-dom": "^18.2.22",
     "@types/react-router-dom": "^5.3.3",
@@ -54,8 +54,8 @@
   "devDependencies": {
     "@preact/compat": "^17.1.2",
     "@preact/preset-vite": "^2.8.2",
-    "@typescript-eslint/eslint-plugin": "^7.3.0",
-    "@typescript-eslint/parser": "^7.3.0",
+    "@typescript-eslint/eslint-plugin": "^7.3.1",
+    "@typescript-eslint/parser": "^7.3.1",
     "concurrently": "^8.2.2",
     "eslint": "^8.57.0",
     "eslint-config-prettier": "^9.1.0",
