+++ interface/src/framework/system/ESPSystemStatus.tsx
@@ -1,5 +1,4 @@
 import AppsIcon from '@mui/icons-material/Apps';
-import BuildIcon from '@mui/icons-material/Build';
 import DeveloperBoardIcon from '@mui/icons-material/DeveloperBoard';
 import DevicesIcon from '@mui/icons-material/Devices';
 import FolderIcon from '@mui/icons-material/Folder';
@@ -7,24 +6,25 @@ import MemoryIcon from '@mui/icons-material/Memory';
 import RefreshIcon from '@mui/icons-material/Refresh';
 import SdCardAlertIcon from '@mui/icons-material/SdCardAlert';
 import SdStorageIcon from '@mui/icons-material/SdStorage';
-import TimerIcon from '@mui/icons-material/Timer';
 import { Avatar, Box, Button, Divider, List, ListItem, ListItemAvatar, ListItemText } from '@mui/material';
 
 import { useRequest } from 'alova';
 import type { FC } from 'react';
 
 import * as SystemApi from 'api/system';
-import { ButtonRow, FormLoader, SectionContent } from 'components';
+import { ButtonRow, FormLoader, SectionContent, useLayoutTitle } from 'components';
 import { useI18nContext } from 'i18n/i18n-react';
 
 function formatNumber(num: number) {
   return new Intl.NumberFormat().format(num);
 }
 
-const SystemStatusForm: FC = () => {
+const ESPSystemStatus: FC = () => {
   const { LL } = useI18nContext();
 
-  const { data: data, send: loadData, error } = useRequest(SystemApi.readSystemStatus, { force: true });
+  useLayoutTitle(LL.STATUS_OF('ESP32'));
+
+  const { data: data, send: loadData, error } = useRequest(SystemApi.readESPSystemStatus, { force: true });
 
   const content = () => {
     if (!data) {
@@ -34,24 +34,6 @@ const SystemStatusForm: FC = () => {
     return (
       <>
         <List>
-          <ListItem>
-            <ListItemAvatar>
-              <Avatar sx={{ bgcolor: '#5f9a5f', color: 'white' }}>
-                <BuildIcon />
-              </Avatar>
-            </ListItemAvatar>
-            <ListItemText primary={LL.EMS_ESP_VER()} secondary={data.emsesp_version} />
-          </ListItem>
-          <Divider variant="inset" component="li" />
-          <ListItem>
-            <ListItemAvatar>
-              <Avatar sx={{ bgcolor: '#5f9a5f', color: 'white' }}>
-                <TimerIcon />
-              </Avatar>
-            </ListItemAvatar>
-            <ListItemText primary={LL.UPTIME()} secondary={data.uptime} />
-          </ListItem>
-          <Divider variant="inset" component="li" />
           <ListItem>
             <ListItemAvatar>
               <Avatar sx={{ bgcolor: '#5f9a5f', color: 'white' }}>
@@ -151,7 +133,6 @@ const SystemStatusForm: FC = () => {
               secondary={formatNumber(data.fs_used) + ' KB / ' + formatNumber(data.fs_free) + ' KB'}
             />
           </ListItem>
-          <Divider variant="inset" component="li" />
         </List>
         <Box display="flex" flexWrap="wrap">
           <Box flexGrow={1} sx={{ '& button': { mt: 2 } }}>
@@ -166,7 +147,7 @@ const SystemStatusForm: FC = () => {
     );
   };
 
-  return <SectionContent title={LL.STATUS_OF(LL.SYSTEM(1))}>{content()}</SectionContent>;
+  return <SectionContent>{content()}</SectionContent>;
 };
 
-export default SystemStatusForm;
+export default ESPSystemStatus;
