+++ interface/src/framework/system/Status.tsx
@@ -1,13 +1,13 @@
 import { Tab } from '@mui/material';
 import { Navigate, Routes, Route } from 'react-router-dom';
 import SystemLog from './SystemLog';
-import SystemStatusForm from './SystemStatusForm';
+import SystemStatus from './SystemStatus';
 import type { FC } from 'react';
 
 import { useRouterTab, RouterTabs, useLayoutTitle } from 'components';
 
 import { useI18nContext } from 'i18n/i18n-react';
-import EMSStatus from 'project/EMSStatus';
+import Activity from 'project/Activity';
 
 const Status: FC = () => {
   const { LL } = useI18nContext();
@@ -20,12 +20,12 @@ const Status: FC = () => {
     <>
       <RouterTabs value={routerTab}>
         <Tab value="status" label={LL.SYSTEM(1)} />
-        <Tab value="emsesp-status" label="EMS-ESP" />
-        <Tab value="log" label={LL.LOG_OF(LL.SYSTEM(2))} />
+        <Tab value="activity" label={LL.ACTIVITY()} />
+        <Tab value="log" label={LL.LOG_OF('')} />
       </RouterTabs>
       <Routes>
-        <Route path="status" element={<SystemStatusForm />} />
-        <Route path="emsesp-status" element={<EMSStatus />} />
+        <Route path="status" element={<SystemStatus />} />
+        <Route path="activity" element={<Activity />} />
         <Route path="log" element={<SystemLog />} />
         <Route path="*" element={<Navigate replace to="status" />} />
       </Routes>
