--- interface/src/project/EMSStatus.tsx
@@ -1,279 +1,0 @@
-import CancelIcon from '@mui/icons-material/Cancel';
-import DeviceHubIcon from '@mui/icons-material/DeviceHub';
-import DirectionsBusIcon from '@mui/icons-material/DirectionsBus';
-import PermScanWifiIcon from '@mui/icons-material/PermScanWifi';
-import RefreshIcon from '@mui/icons-material/Refresh';
-import {
-  Avatar,
-  Box,
-  Button,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogTitle,
-  List,
-  ListItem,
-  ListItemAvatar,
-  ListItemText,
-  useTheme
-} from '@mui/material';
-import { Body, Cell, Header, HeaderCell, HeaderRow, Row, Table } from '@table-library/react-table-library/table';
-import { useTheme as tableTheme } from '@table-library/react-table-library/theme';
-import { useRequest } from 'alova';
-import { useContext, useEffect, useState } from 'react';
-import { toast } from 'react-toastify';
-
-import * as EMSESP from './api';
-import { busConnectionStatus } from './types';
-import type { Stat, Status } from './types';
-import type { Theme } from '@mui/material';
-
-import type { Translation } from 'i18n/i18n-types';
-import type { FC } from 'react';
-import { dialogStyle } from 'CustomTheme';
-import { ButtonRow, FormLoader, SectionContent } from 'components';
-import { AuthenticatedContext } from 'contexts/authentication';
-import { useI18nContext } from 'i18n/i18n-react';
-
-export const isConnected = ({ status }: Status) => status !== busConnectionStatus.BUS_STATUS_OFFLINE;
-
-const busStatusHighlight = ({ status }: Status, theme: Theme) => {
-  switch (status) {
-    case busConnectionStatus.BUS_STATUS_TX_ERRORS:
-      return theme.palette.warning.main;
-    case busConnectionStatus.BUS_STATUS_CONNECTED:
-      return theme.palette.success.main;
-    case busConnectionStatus.BUS_STATUS_OFFLINE:
-      return theme.palette.error.main;
-    default:
-      return theme.palette.warning.main;
-  }
-};
-
-const showQuality = (stat: Stat) => {
-  if (stat.q === 0 || stat.s + stat.f === 0) {
-    return;
-  }
-  if (stat.q === 100) {
-    return <div style={{ color: '#00FF7F' }}>{stat.q}%</div>;
-  }
-  if (stat.q >= 95) {
-    return <div style={{ color: 'orange' }}>{stat.q}%</div>;
-  } else {
-    return <div style={{ color: 'red' }}>{stat.q}%</div>;
-  }
-};
-
-const EMSStatus: FC = () => {
-  const { data: data, send: loadData, error } = useRequest(EMSESP.readStatus);
-
-  const { LL } = useI18nContext();
-
-  const theme = useTheme();
-  const [confirmScan, setConfirmScan] = useState<boolean>(false);
-
-  const { me } = useContext(AuthenticatedContext);
-
-  const { send: scanDevices } = useRequest(EMSESP.scanDevices, {
-    immediate: false
-  });
-
-  const stats_theme = tableTheme({
-    Table: `
-      --data-table-library_grid-template-columns: repeat(1, minmax(0, 1fr)) 90px 90px 80px;
-    `,
-    BaseRow: `
-      font-size: 14px;
-    `,
-    HeaderRow: `
-      text-transform: uppercase;
-      background-color: black;
-      color: #90CAF9;
-
-      .th {
-        height: 36px;
-        border-bottom: 1px solid #565656;
-      }
-    `,
-    Row: `
-      .td {
-        padding: 8px;
-        border-top: 1px solid #565656;
-        border-bottom: 1px solid #565656;
-      }
-
-      &:nth-of-type(odd) .td {
-        background-color: #303030;
-      }
-      &:nth-of-type(even) .td {
-        background-color: #1e1e1e;
-      }
-    `,
-    BaseCell: `
-      &:not(:first-of-type) {
-        text-align: center;
-      }
-    `
-  });
-
-  useEffect(() => {
-    const timer = setInterval(() => loadData(), 30000);
-    return () => {
-      clearInterval(timer);
-    };
-  });
-
-  const showName = (id: any) => {
-    const name: keyof Translation['STATUS_NAMES'] = id;
-    return LL.STATUS_NAMES[name]();
-  };
-
-  const formatDurationSec = (duration_sec: number) => {
-    const days = Math.trunc((duration_sec * 1000) / 86400000);
-    const hours = Math.trunc((duration_sec * 1000) / 3600000) % 24;
-    const minutes = Math.trunc((duration_sec * 1000) / 60000) % 60;
-    const seconds = Math.trunc((duration_sec * 1000) / 1000) % 60;
-
-    let formatted = '';
-    if (days) {
-      formatted += LL.NUM_DAYS({ num: days }) + ' ';
-    }
-    if (hours) {
-      formatted += LL.NUM_HOURS({ num: hours }) + ' ';
-    }
-    if (minutes) {
-      formatted += LL.NUM_MINUTES({ num: minutes }) + ' ';
-    }
-    formatted += LL.NUM_SECONDS({ num: seconds });
-    return formatted;
-  };
-
-  const busStatus = () => {
-    if (data) {
-      switch (data.status) {
-        case busConnectionStatus.BUS_STATUS_CONNECTED:
-          return LL.CONNECTED(0) + ' (' + formatDurationSec(data.uptime) + ')';
-        case busConnectionStatus.BUS_STATUS_TX_ERRORS:
-          return LL.TX_ISSUES();
-        case busConnectionStatus.BUS_STATUS_OFFLINE:
-          return LL.DISCONNECTED();
-      }
-    }
-    return 'Unknown';
-  };
-
-  const scan = async () => {
-    await scanDevices()
-      .then(() => {
-        toast.info(LL.SCANNING() + '...');
-      })
-      .catch((err) => {
-        toast.error(err.message);
-      });
-    setConfirmScan(false);
-  };
-
-  const renderScanDialog = () => (
-    <Dialog sx={dialogStyle} open={confirmScan} onClose={() => setConfirmScan(false)}>
-      <DialogTitle>{LL.SCAN_DEVICES()}</DialogTitle>
-      <DialogContent dividers>{LL.EMS_SCAN()}</DialogContent>
-      <DialogActions>
-        <Button startIcon={<CancelIcon />} variant="outlined" onClick={() => setConfirmScan(false)} color="secondary">
-          {LL.CANCEL()}
-        </Button>
-        <Button startIcon={<PermScanWifiIcon />} variant="outlined" onClick={scan} color="primary">
-          {LL.SCAN()}
-        </Button>
-      </DialogActions>
-    </Dialog>
-  );
-
-  const content = () => {
-    if (!data) {
-      return <FormLoader onRetry={loadData} errorMessage={error?.message} />;
-    }
-
-    return (
-      <>
-        <List>
-          <ListItem>
-            <ListItemAvatar>
-              <Avatar sx={{ bgcolor: busStatusHighlight(data, theme) }}>
-                <DirectionsBusIcon />
-              </Avatar>
-            </ListItemAvatar>
-            <ListItemText primary={LL.EMS_BUS_STATUS()} secondary={busStatus()} />
-          </ListItem>
-          <ListItem>
-            <ListItemAvatar>
-              <Avatar sx={{ bgcolor: theme.palette.success.main }}>
-                <DeviceHubIcon />
-              </Avatar>
-            </ListItemAvatar>
-            <ListItemText
-              primary={LL.ACTIVE_DEVICES()}
-              secondary={
-                LL.NUM_DEVICES({ num: data.num_devices }) +
-                ', ' +
-                LL.NUM_TEMP_SENSORS({ num: data.num_sensors }) +
-                ', ' +
-                LL.NUM_ANALOG_SENSORS({ num: data.num_analogs })
-              }
-            />
-          </ListItem>
-          <Box m={3} />
-          <Table data={{ nodes: data.stats }} theme={stats_theme} layout={{ custom: true }}>
-            {(tableList: any) => (
-              <>
-                <Header>
-                  <HeaderRow>
-                    <HeaderCell resize />
-                    <HeaderCell stiff>{LL.SUCCESS()}</HeaderCell>
-                    <HeaderCell stiff>{LL.FAIL()}</HeaderCell>
-                    <HeaderCell stiff>{LL.QUALITY()}</HeaderCell>
-                  </HeaderRow>
-                </Header>
-                <Body>
-                  {tableList.map((stat: Stat) => (
-                    <Row key={stat.id} item={stat}>
-                      <Cell>{showName(stat.id)}</Cell>
-                      <Cell stiff>{Intl.NumberFormat().format(stat.s)}</Cell>
-                      <Cell stiff>{Intl.NumberFormat().format(stat.f)}</Cell>
-                      <Cell stiff>{showQuality(stat)}</Cell>
-                    </Row>
-                  ))}
-                </Body>
-              </>
-            )}
-          </Table>
-        </List>
-        {renderScanDialog()}
-        <Box display="flex" flexWrap="wrap">
-          <Box flexGrow={1} sx={{ '& button': { mt: 2 } }}>
-            <Button startIcon={<RefreshIcon />} variant="outlined" color="secondary" onClick={loadData}>
-              {LL.REFRESH()}
-            </Button>
-          </Box>
-          {me.admin && (
-            <Box flexWrap="nowrap" whiteSpace="nowrap">
-              <ButtonRow>
-                <Button
-                  startIcon={<PermScanWifiIcon />}
-                  variant="outlined"
-                  color="primary"
-                  onClick={() => setConfirmScan(true)}
-                >
-                  {LL.SCAN_DEVICES()}
-                </Button>
-              </ButtonRow>
-            </Box>
-          )}
-        </Box>
-      </>
-    );
-  };
-
-  return <SectionContent title={LL.EMS_BUS_STATUS_TITLE()}>{content()}</SectionContent>;
-};
-
-export default EMSStatus;
