+++ interface/src/types/system.ts
@@ -1,5 +1,6 @@
-export interface SystemStatus {
-  // TODO fix this next
+import type { busConnectionStatus } from 'project/types';
+
+export interface ESPSystemStatus {
   emsesp_version: string;
   esp_platform: string;
   max_alloc_heap: number;
@@ -17,11 +18,21 @@ export interface SystemStatus {
   app_free: number;
   fs_used: number;
   fs_free: number;
-  uptime: string;
   free_mem: number;
   psram_size?: number;
   free_psram?: number;
   has_loader: boolean;
+}
+
+export interface SystemStatus {
+  emsesp_version: string;
+  esp_platform: string;
+  status: busConnectionStatus;
+  tx_mode: number;
+  uptime: number;
+  num_devices: number;
+  num_sensors: number;
+  num_analogs: number;
 }
 
 export interface OTASettings {
